redirection vers supprimerPro.php, listeClients.php ou listeCommandes.php
Redirection sur les zone de suppression respectives

+++ administration/requetesSupprimer.php
@@ -8,18 +8,20 @@ if(isset($_POST['identifiant'])&&isset($_POST['nomProduit'])){
 	$nomProduit=verifierDonne($_POST['nomProduit']);
 	$bdd->requetes('DELETE FROM produits WHERE id_pro=? AND nom_pro=?',
 				array($identifiant, $nomProduit));
-		
+	header("Location:adminHeader.php?admin=supprimerPro");
 		
 		//Suprimr clients
 }elseif(isset($_POST['suppCli'])){
 	$idCli=verifierDonne($_POST['suppCli']);
 	$bdd->requetes('DELETE FROM clients WHERE id_cli=?',
 			array($idCli));
+	header("Location:adminHeader.php?admin=listeClients");
 			
-}else{ //Supprimer Commande
+}elseif(isset($_POST['suppCom'])){ //Supprimer Commande
 	$idCom=verifierDonne($_POST['suppCom']);
 	$bdd->requetes('DELETE FROM commandes WHERE id_com=?',
 				array($idCom));
-}
-			
+	header("Location:adminHeader.php?admin=listeCommandes");
+}else{	
 	header("Location:adminHeader.php");
+	}
